requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
psycopg2-binary
pandas
requests
bs4